SpeechToText/requirements.txt
# packages in environment at C:\Users\dream\anaconda3\envs\speach_to_text:
#
# Name                    Version                   Build  Channel
aiohttp                   3.8.3                    pypi_0    pypi
aiosignal                 1.3.1                    pypi_0    pypi
async-timeout             4.0.2                    pypi_0    pypi
attrs                     22.2.0                   pypi_0    pypi
autopep8                  2.0.1                    pypi_0    pypi
blas                      1.0                         mkl  
brotlipy                  0.7.0           py39h2bbff1b_1003  
ca-certificates           2023.01.10           haa95532_0  
cachetools                5.3.0                    pypi_0    pypi
certifi                   2022.12.7        py39haa95532_0  
cffi                      1.15.1           py39h2bbff1b_3  
charset-normalizer        2.1.1                    pypi_0    pypi
click                     8.1.3                    pypi_0    pypi
colorama                  0.4.6                    pypi_0    pypi
cpuonly                   2.0                           0    pytorch
cryptography              38.0.4           py39h21b164f_0  
filelock                  3.9.0                    pypi_0    pypi
flit-core                 3.6.0              pyhd3eb1b0_0  
freetype                  2.12.1               ha860e81_0  
frozenlist                1.3.3                    pypi_0    pypi
google-api-core           2.11.0                   pypi_0    pypi
google-api-python-client  2.74.0                   pypi_0    pypi
google-auth               2.16.0                   pypi_0    pypi
google-auth-httplib2      0.1.0                    pypi_0    pypi
google-auth-oauthlib      0.8.0                    pypi_0    pypi
google-cloud-speech       2.17.1                   pypi_0    pypi
googleapis-common-protos  1.58.0                   pypi_0    pypi
grpcio                    1.51.1                   pypi_0    pypi
grpcio-status             1.51.1                   pypi_0    pypi
httplib2                  0.21.0                   pypi_0    pypi
huggingface-hub           0.12.0                   pypi_0    pypi
idna                      3.4              py39haa95532_0  
intel-openmp              2021.4.0          haa95532_3556  
joblib                    1.2.0                    pypi_0    pypi
jpeg                      9e                   h2bbff1b_0  
lerc                      3.0                  hd77b12b_0  
libdeflate                1.8                  h2bbff1b_5  
libpng                    1.6.37               h2a8f88b_0  
libtiff                   4.5.0                h8a3f274_0  
libuv                     1.40.0               he774522_0  
libwebp                   1.2.4                h2bbff1b_0  
libwebp-base              1.2.4                h2bbff1b_0  
lz4-c                     1.9.4                h2bbff1b_0  
mkl                       2021.4.0           haa95532_640  
mkl-service               2.4.0            py39h2bbff1b_0  
mkl_fft                   1.3.1            py39h277e83a_0  
mkl_random                1.2.2            py39hf11a4ad_0  
multidict                 6.0.4                    pypi_0    pypi
nltk                      3.8.1                    pypi_0    pypi
numpy                     1.24.1                   pypi_0    pypi
numpy-base                1.23.5           py39h4da318b_0  
oauthlib                  3.2.2                    pypi_0    pypi
openai                    0.26.4                   pypi_0    pypi
openssl                   1.1.1s               h2bbff1b_0  
packaging                 23.0                     pypi_0    pypi
pillow                    9.3.0            py39hdc2b20a_1  
pip                       22.3.1           py39haa95532_0  
proto-plus                1.22.2                   pypi_0    pypi
protobuf                  4.21.12                  pypi_0    pypi
pyasn1                    0.4.8                    pypi_0    pypi
pyasn1-modules            0.2.8                    pypi_0    pypi
pyaudio                   0.2.11                   pypi_0    pypi
pycodestyle               2.10.0                   pypi_0    pypi
pycparser                 2.21               pyhd3eb1b0_0  
pyopenssl                 22.0.0             pyhd3eb1b0_0  
pyparsing                 3.0.9                    pypi_0    pypi
pysocks                   1.7.1            py39haa95532_0  
python                    3.9.16               h6244533_0  
pytorch                   1.13.1              py3.9_cpu_0    pytorch
pytorch-mutex             1.0                         cpu    pytorch
pyyaml                    6.0                      pypi_0    pypi
regex                     2022.10.31               pypi_0    pypi
requests                  2.28.2                   pypi_0    pypi
requests-oauthlib         1.3.1                    pypi_0    pypi
rsa                       4.9                      pypi_0    pypi
scikit-learn              1.2.1                    pypi_0    pypi
scipy                     1.10.0                   pypi_0    pypi
sentence-transformers     2.2.2                    pypi_0    pypi
sentencepiece             0.1.97                   pypi_0    pypi
setuptools                65.6.3           py39haa95532_0  
six                       1.16.0             pyhd3eb1b0_1  
sqlite                    3.40.1               h2bbff1b_0  
threadpoolctl             3.1.0                    pypi_0    pypi
tk                        8.6.12               h2bbff1b_0  
tokenizers                0.13.2                   pypi_0    pypi
tomli                     2.0.1                    pypi_0    pypi
torchaudio                0.13.1                 py39_cpu    pytorch
torchvision               0.14.1                 py39_cpu    pytorch
tqdm                      4.64.1                   pypi_0    pypi
transformers              4.26.0                   pypi_0    pypi
typing_extensions         4.4.0            py39haa95532_0  
tzdata                    2022g                h04d1e81_0  
uritemplate               4.1.1                    pypi_0    pypi
urllib3                   1.26.14          py39haa95532_0  
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
wheel                     0.37.1             pyhd3eb1b0_0  
win_inet_pton             1.1.0            py39haa95532_0  
wincertstore              0.2              py39haa95532_2  
xz                        5.2.10               h8cc25b3_1  
yarl                      1.8.2                    pypi_0    pypi
zlib                      1.2.13               h8cc25b3_0  
zstd                      1.5.2                h19a0ad4_0  
